all components created and app.jsx seperated. Css also seperated into correct style files
