+++ src/styles/App.css
@@ -8,139 +8,7 @@
     180px 1fr;
 }
 
-.app > .header {
-  grid-area: header;
 
-  display: grid;
-  grid-template-columns: 180px 1fr;
 
-  border-bottom: solid 1px #00000020;
-}
 
-.app > .header > .left-menu {
-  display: grid;
-  grid-auto-flow: column;
-  align-items: center;
-  justify-content: left;
-}
 
-.app > .header > .left-menu .menu-icon {
-  width: 25px;
-  height: 25px;
-  margin: 10px;
-}
-
-.app > .header > .search {
-  display: grid;
-  padding: 6px;
-}
-
-.app > .header > .search .search-bar {
-  border: none;
-  border-radius: 5px;
-  background-color: rgb(240, 240, 240);
-  padding: 0 10px;
-  font-size: 1rem;
-}
-
-.app > .emails {
-  grid-area: emails;
-  padding: 10px 0;
-}
-
-.app > .emails > ul {
-  list-style: none;
-  display: grid;
-}
-
-.app > .left-menu {
-  grid-area: left-menu;
-  padding: 10px;
-  padding-left: 0;
-}
-
-.app > .left-menu .inbox-list {
-  display: grid;
-  grid-gap: 10px;
-}
-
-.app > .left-menu .inbox-list .item {
-  padding: 5px 10px;
-  display: grid;
-  grid-auto-flow: column;
-  justify-content: space-between;
-  align-items: center;
-
-  border-top-right-radius: 20px;
-  border-bottom-right-radius: 20px;
-}
-
-.app > .left-menu .inbox-list .item.active {
-  background-color: #ff000050;
-}
-
-.app > .left-menu .inbox-list .item.toggle {
-  background-color: transparent;
-}
-
-.email {
-  display: grid;
-  grid-template-columns: 30px 30px 150px 1fr;
-  padding: 10px;
-  background-color: white;
-  font-size: 0.9rem;
-
-  border-bottom: solid 1px #00000020;
-}
-
-.email:first-child {
-  border-top: solid 1px #00000020;
-}
-
-.email:hover {
-  box-shadow: -1px 1px 2px 0 #00000040;
-  z-index: 1;
-}
-
-.email.read {
-  background-color: rgb(240, 240, 240);
-}
-
-.email.unread {
-  font-weight: 600;
-}
-
-.email .title {
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.email .star-checkbox {
-  appearance: none;
-  position: relative;
-  display: grid;
-  align-content: center;
-}
-
-.email .star-checkbox:focus {
-  outline: none;
-}
-
-.email .star-checkbox:checked::before,
-.email .star-checkbox::before {
-  content: '';
-  background-size: cover;
-  width: 20px;
-  height: 20px;
-  position: absolute;
-  top: -3px;
-}
-
-.email .star-checkbox::before {
-  background-image: url(https://www.gstatic.com/images/icons/material/system/2x/star_border_black_20dp.png);
-}
-
-.email .star-checkbox:checked::before {
-  background-image: url(https://www.gstatic.com/images/icons/material/system/2x/star_googyellow500_20dp.png);
-}
