+++ src/styles/Email.css
@@ -1,0 +1,9 @@
+.app > .emails {
+    grid-area: emails;
+    padding: 10px 0;
+  }
+  
+  .app > .emails > ul {
+    list-style: none;
+    display: grid;
+  }
