+++ src/styles/Emails.css
@@ -1,0 +1,61 @@
+.email {
+    display: grid;
+    grid-template-columns: 30px 30px 150px 1fr;
+    padding: 10px;
+    background-color: white;
+    font-size: 0.9rem;
+  
+    border-bottom: solid 1px #00000020;
+  }
+  
+  .email:first-child {
+    border-top: solid 1px #00000020;
+  }
+  
+  .email:hover {
+    box-shadow: -1px 1px 2px 0 #00000040;
+    z-index: 1;
+  }
+  
+  .email.read {
+    background-color: rgb(240, 240, 240);
+  }
+  
+  .email.unread {
+    font-weight: 600;
+  }
+  
+  .email .title {
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+  }
+  
+  .email .star-checkbox {
+    appearance: none;
+    position: relative;
+    display: grid;
+    align-content: center;
+  }
+  
+  .email .star-checkbox:focus {
+    outline: none;
+  }
+  
+  .email .star-checkbox:checked::before,
+  .email .star-checkbox::before {
+    content: '';
+    background-size: cover;
+    width: 20px;
+    height: 20px;
+    position: absolute;
+    top: -3px;
+  }
+  
+  .email .star-checkbox::before {
+    background-image: url(https://www.gstatic.com/images/icons/material/system/2x/star_border_black_20dp.png);
+  }
+  
+  .email .star-checkbox:checked::before {
+    background-image: url(https://www.gstatic.com/images/icons/material/system/2x/star_googyellow500_20dp.png);
+  }
