+++ src/styles/Header.css
@@ -1,0 +1,34 @@
+.app > .header {
+    grid-area: header;
+  
+    display: grid;
+    grid-template-columns: 180px 1fr;
+  
+    border-bottom: solid 1px #00000020;
+  }
+  
+  .app > .header > .left-menu {
+    display: grid;
+    grid-auto-flow: column;
+    align-items: center;
+    justify-content: left;
+  }
+  
+  .app > .header > .left-menu .menu-icon {
+    width: 25px;
+    height: 25px;
+    margin: 10px;
+  }
+  
+  .app > .header > .search {
+    display: grid;
+    padding: 6px;
+  }
+  
+  .app > .header > .search .search-bar {
+    border: none;
+    border-radius: 5px;
+    background-color: rgb(240, 240, 240);
+    padding: 0 10px;
+    font-size: 1rem;
+  }
