+++ src/styles/LeftMenu.css
@@ -1,0 +1,29 @@
+.app > .left-menu {
+    grid-area: left-menu;
+    padding: 10px;
+    padding-left: 0;
+  }
+  
+  .app > .left-menu .inbox-list {
+    display: grid;
+    grid-gap: 10px;
+  }
+  
+  .app > .left-menu .inbox-list .item {
+    padding: 5px 10px;
+    display: grid;
+    grid-auto-flow: column;
+    justify-content: space-between;
+    align-items: center;
+  
+    border-top-right-radius: 20px;
+    border-bottom-right-radius: 20px;
+  }
+  
+  .app > .left-menu .inbox-list .item.active {
+    background-color: #ff000050;
+  }
+  
+  .app > .left-menu .inbox-list .item.toggle {
+    background-color: transparent;
+  }
